dwc_eqos - fix verifier complaint about interrupt lock (#30)
Interrupt lock can't be acquired until interrupts are connected, so
verifier complains about attempts to acquire/release the interrupt lock
in D0Entry/D0Exit. Fix by doing the work outside the lock, which is ok
because we don't really need a lock for this during D0Entry/D0Exit.