Rebase INF files against 22621
The .sys files are patched versions of the drivers from Windows 22621.

The .inf files that I added are patched versions of the INF files from
27557, which doesn't match the .sys files.

Fix by re-doing the .inf file patch against the 22621 INF files. Not
much difference, and maybe it doesn't matter, but seems like the right
thing to do.