Add Synopsys DWC MSHC driver
This is an sdport-based miniport driver for the two MSHC instances on
RK3588: one for SD card slot, the other for SDIO Wi-Fi.

Features:
- Speed modes: Default, HS, SDR50, SDR104
- Internal 32-bit scatter/gather DMA support
- Card detection
- Crashdump support

To-do:
- Test eMMC support
- SDIO (e.g. CMD53)
- Support SD version 1 cards (they don't respond to CMD8 by design,
which makes sdport assume it's not an SD card at all)

Known issues:
- Plugging/unplugging the card quickly can lead to a bugcheck. This
happens due to race conditions in sdport w.r.t. card removal and
request completion. It ends up dereferencing a NULL pointer, which was
supposed to be the outstanding request.

Signed-off-by: Mario Bălănică <mariobalanica02@gmail.com>