﻿dwcmshc: Fix signaling voltage mismatch
SD cards can only be switched back to 3.3V after a power cycle.

The firmware does not currently implement card power control (and some
boards don't even have a switch to begin with), so we need to ensure
that the card is not initialized before being able to safely change
signaling on the host side. This should only happen after the card is
unplugged.

Signed-off-by: Mario Bălănică <mariobalanica02@gmail.com>